+++ scraping/reddit_scrape.py
@@ -4,36 +4,35 @@ reddit_scrape.py
 Part of the "Instant Serotonin" project by OperaVaria.
 """
 
-# External module imports:
+# Imports:
+from sys import platform
 import praw
-import pickle
-
-# Login variable imports:
-from reddit_login import id, secret
+from scraping.post_class import Post
 
 
-def reddit_collect(sub_red_name, user_agent_name, file_name):
+def reddit_collect(sub_red_name, keys):
     """Reddit scraper function.
        Gets all pictures and gifs from hot top 30 of any subreddit.
-       Pickles list into a binary file for quick reading.
-       Prints number of items collected.
-       PRAW and personal reddit app used."""    
+       PRAW and personal reddit app login used."""
     # Login details.
     reddit = praw.Reddit(
-        client_id = id,
-        client_secret = secret,
-        user_agent = user_agent_name)    
+        client_id = keys["reddit_app"]["id"],
+        client_secret = keys["reddit_app"]["secret"],
+        user_agent = f"{platform}:instant-serotonin:v2.0.0 (by u/just_want_api)")
     # Set subreddit(s).
     subreddit = reddit.subreddit(sub_red_name)
-    # Scrape posts, create submission list.
-    reddit_sub_list = [submission for submission in subreddit.hot(limit=30)
-                       if submission.url.lower().endswith((".png", ".jpg", ".jpeg", ".gif"))]    
-    # Dump list to picke file.
-    with open(file_name, "wb") as file:
-        pickle.dump(reddit_sub_list, file)    
-    # Print number of submissions in set.
-    print("Number of submissions: " + str(len(reddit_sub_list)))
-    return
+    # Declare post list.
+    reddit_post_list = []
+    # Scrape submissions, create Post objects, append to list.
+    for submission in subreddit.hot(limit=30):
+        if submission.url.lower().endswith((".png", ".jpg", ".jpeg", ".gif")):
+            post = Post(submission.title, f"r/{submission.subreddit.display_name}", submission.author,
+                        f"https://www.reddit.com/r/{submission.subreddit.display_name}/",
+                        f"https://www.reddit.com/user/{submission.author}/",
+                        f"https://www.reddit.com{submission.permalink}", submission.url)
+            reddit_post_list.append(post)
+    # Return list.
+    return reddit_post_list
 
 
 # Print on accidental run: